Revert "My assignment component"
