--- src/css/MyAssignment.css
@@ -1,69 +1,0 @@
-.assignmentContainer {
-    max-width: 600px;
-    margin: 20px auto;
-    padding: 20px;
-    border-radius: 8px;
-    background-color: #f9f9f9;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    font-family: 'Arial', sans-serif;
-}
-
-button {
-    background-color: #4CAF50;
-    /* Green */
-    color: white;
-    padding: 10px 15px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-button:hover {
-    background-color: #45a049;
-}
-
-label {
-    display: block;
-    margin: 10px 0 5px;
-    font-weight: bold;
-}
-
-input[type="text"] {
-    width: 100%;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    box-sizing: border-box;
-    transition: border-color 0.3s;
-}
-
-input[type="text"]:focus {
-    border-color: #4CAF50;
-    outline: none;
-}
-
-.selectedWordsContainer {
-    margin-top: 20px;
-}
-
-.words {
-    display: inline-block;
-    background-color: #e7f3fe;
-    color: #010c11;
-    padding: 5px 10px;
-    border-radius: 4px;
-    margin: 5px;
-    font-size: 14px;
-}
-
-.title,
-.description {
-    margin: 10px 0;
-    font-size: 16px;
-    color: #1d0202;
-}
-
-.assignmentIfo{
-    display: flex;
-}
